MC: Fix the condvar reversible race & let the exploration handle
In some cases, we say that a race is reversible but the truth is that
it isn't. In those cases, we can either implement some computation
to verify before hand, or (what we do for now) we let the exploration
discover by itself that it can't execute what it wants. This has the
default of artificialy boosting the number of visited traces.