[MC] Add a missing case for reversible race
This one was tricky: a condvar_async_lock releases a mutex. So it is
dependent with a mutex wait. But since mutex_wait has a smaller type,
it was not dealt with in the condvar_async_lock::depends method.
But reversible_race() is not symmetric. This is how I missed that in
the first place.