Merge ActorMain and Actor
This is somewhat suboptimal because the actors need to be manually run
with start() after their creation, because I have a race condition
between the creator actor that runs the constructor of the child, and
the created one that runs its run() function.

It may happen that the ctor is still ongoing when the child kicks in...

This is still TBC as ActorMain shall now die, and maybe the previous
issue can be fixed with some synchronized keywords.