Java: Solve a race condition
There were a race condition between:
 - the actor ctor in parent thread
 - the actor run() in actor's thread
where the run() could sometimes execute before the end of its ctor.

This is solved simply by requesting the actor thread to only start
executing in the next simgrid scheduling round. Yup, that's a stupid
little yield() just before the start of the created actor. I think
it's sufficient, and the (rare) existing test seem to say so too.

This is better that the previous solution, as there is no need for a
start() in user code.