Make a several dag tests pass in Java
Kill some SWIG dumb types on the way, and kill quite a lot of SWIG
lines in the C++ code. I'm soon done with this. I hope, at least.

Continue implementing the bindings, even if some currently untested
things are known to fail (thanks to the tests that are not activated)