Do not use inline static in templated class to avoid a clang bug revealed by Java
My bug is very similar to https://github.com/llvm/llvm-project/issues/80365
but in my case, it works with g++ 14.8 and fails with clang++ 19.1.5
(both from Debian). Or, more precisely, it works most of the time but
when the resulting lib is loaded by the JVM as a plugin, the static
inline fields seem to have several addresses, as if it were loaded in
several locations.

Anyway, this many template instanciations are ugly but at least it
works.