MC: use SOCK_STREAM even on platforms that have AF_UNIX + SOCK_SEQPACKET
Better bufferization now that our channel fully supports it, and less
portability nightmare.

On `mcmini-barber_shop_ok 5 2 0 0` this reduces the amount of recvfrom
from 989k down to 515k. The total amount of syscalls goes from 1.56M down
to 1.05M.

The wallclock is marginally reduced (1:14 down to 1:11) but it's hard
to say for sure from a single benchmark because this job only gets
about 20% of the time. I guess we should optimize on the app side now.