Strenghten our memory consistency model around the refcounting
The comment in ActorImpl was right: we really need to go for the
acq_rel consistency model here, as neither acquire nor release was
enough.

It was never tested properly because simgrid rarely runs in //. It
fails in the parallel explorer that we are working on for the MC, so
let's solve this issue for everybody.